refactor: merged Map component to +page.svelte since it is not used any other places.
