+++ .changeset/quiet-kiwis-protect.md
@@ -1,0 +1,5 @@
+---
+"geohub": patch
+---
+
+refactor: merged Map component to +page.svelte since it is not used any other places.
